% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmoutlier.R
\name{mmoutlier}
\alias{mmoutlier}
\title{Identifying outliers from univariate data and replace the outliers by max-min Method'}
\usage{
mmoutlier(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Identifying outliers from univariate data and replace the outliers by max-min Method'
}
